Change nSpendHeight and GetSpendHeight to int64_t (fixes H2)
int nSpendHeight in CheckTxInputs caused signed/unsigned mismatch with
coin.nHeight (uint32_t:31). C++ arithmetic conversions promoted int to
unsigned, so a negative or underflowing nSpendHeight could produce a
huge positive depth, silently bypassing the coinbase/pegout maturity
checks. Changing to int64_t ensures coin.nHeight is promoted upward to
int64_t instead, making underflow impossible and negative values
detectable. GetSpendHeight updated to match; fuzz caller updated to use
ConsumeIntegralInRange<int64_t>. Consistent with the H1 fix that moved
block-height fields in consensus/params.h to int64_t.

+++ src/validation.cpp
@@ -1531,7 +1531,7 @@ bool CScriptCheck::operator()() {
     return VerifyScript(scriptSig, m_tx_out.scriptPubKey, witness, nFlags, CachingTransactionSignatureChecker(ptxTo, nIn, m_tx_out.nValue, cacheStore, *txdata), &error);
 }
 
-int GetSpendHeight(const CCoinsViewCache& inputs)
+int64_t GetSpendHeight(const CCoinsViewCache& inputs)
 {
     LOCK(cs_main);
     CBlockIndex* pindexPrev = LookupBlockIndex(inputs.GetBestBlock());
